Move bootstrap methods out of Registration
